backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.11
django-environ==0.11.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
bcrypt==4.0.1
reportlab

# Needed for djangorestframework-simplejwt in Python 3.12+
setuptools>=70.0.0
